Trait float_eq::AssertFloatEq

pub trait AssertFloatEq<Rhs: ?Sized = Self>: FloatEq<Rhs> {
    type DebugAbsDiff: Debug + Sized + FloatEqDebugUlpsDiff;
    type DebugEpsilon: Debug + FloatEqUlpsEpsilon;
    fn debug_abs_diff(&self, other: &Rhs) -> Self::DebugAbsDiff;

    fn debug_ulps_diff(&self, other: &Rhs) -> DebugUlpsDiff<Self::DebugAbsDiff>;

    fn debug_abs_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::Epsilon
    ) -> Self::DebugEpsilon;

    fn debug_rmax_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::Epsilon
    ) -> Self::DebugEpsilon;

    fn debug_rmin_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::Epsilon
    ) -> Self::DebugEpsilon;

    fn debug_r1st_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::Epsilon
    ) -> Self::DebugEpsilon;

    fn debug_r2nd_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::Epsilon
    ) -> Self::DebugEpsilon;

    fn debug_ulps_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &UlpsEpsilon<Self::Epsilon>
    ) -> UlpsEpsilon<Self::DebugEpsilon>
    where
        UlpsEpsilon<Self::DebugEpsilon>: Sized;

    fn debug_rel_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::Epsilon
    ) -> Self::DebugEpsilon { ... }
}

Debug context for when an assert fails.

This is used internally by the assert_float_eq! family of macros.

To implement this trait over a new type, see How to compare custom types.

Associated Types

type DebugAbsDiff: Debug + Sized + FloatEqDebugUlpsDiff

The absolute difference between two values, displayed to the user via fmt::Debug when an assert fails.

This is usually the wider of Self and Rhs.

type DebugEpsilon: Debug + FloatEqUlpsEpsilon

The per-field epsilon value used for comparison between two values, displayed to the user via fmt::Debug when an assert fails.

This should match Self::Epsilon.

Loading content...
Required methods

fn debug_abs_diff(&self, other: &Rhs) -> Self::DebugAbsDiff

Always positive absolute difference between two values.

Implementations should be the equivalent of:

(self - other).abs()

fn debug_ulps_diff(&self, other: &Rhs) -> DebugUlpsDiff<Self::DebugAbsDiff>

Always positive absolute difference between two values in terms of ULPs.

For primitive values, this should be a partial function that returns:

Some(0) if either argument is 0.0 or -0.0

None if either argument is NaN

None if the arguments have differing signs

Some(bitwise-difference) otherwise

For composite types, this should return per-field recursively calculated results in order to present the most possible context to the user.

Implementations over primitive types should be the equivalent of (using f32 as an example):

if self == other {
    Some(0)
} else if self.is_nan() || other.is_nan() {
    None
} else if self.is_sign_positive() != other.is_sign_positive() {
    None
} else {
    let a = self.to_bits();
    let b = other.to_bits();
    let max = a.max(b);
    let min = a.min(b);
    Some(max - min)
}

fn debug_abs_epsilon(
&self,
other: &Rhs,
max_diff: &Self::Epsilon
) -> Self::DebugEpsilon

The epsilon used by an abs absolute tolerance comparison, displayed when an assert fails.

fn debug_rmax_epsilon(
&self,
other: &Rhs,
max_diff: &Self::Epsilon
) -> Self::DebugEpsilon

The epsilon used by an rmax relative tolerance comparison, displayed when an assert fails.

This should take into account the fact that the epsilon values are scaled based on the size of their inputs.

fn debug_rmin_epsilon(
&self,
other: &Rhs,
max_diff: &Self::Epsilon
) -> Self::DebugEpsilon

The epsilon used by an rmin relative tolerance comparison, displayed when an assert fails.

This should take into account the fact that the epsilon values are scaled based on the size of their inputs.

fn debug_r1st_epsilon(
&self,
other: &Rhs,
max_diff: &Self::Epsilon
) -> Self::DebugEpsilon

The epsilon used by an r1st relative tolerance comparison, displayed when an assert fails.

This should take into account the fact that the epsilon values are scaled based on the size of their inputs.

fn debug_r2nd_epsilon(
&self,
other: &Rhs,
max_diff: &Self::Epsilon
) -> Self::DebugEpsilon

The epsilon used by an r2nd relative tolerance comparison, displayed when an assert fails.

This should take into account the fact that the epsilon values are scaled based on the size of their inputs.

fn debug_ulps_epsilon(
&self,
other: &Rhs,
max_diff: &UlpsEpsilon<Self::Epsilon>
) -> UlpsEpsilon<Self::DebugEpsilon> where
UlpsEpsilon<Self::DebugEpsilon>: Sized,

The epsilon used by an ulps ULPs comparison, displayed when an assert fails.

Loading content...
Provided methods

fn debug_rel_epsilon(
&self,
other: &Rhs,
max_diff: &Self::Epsilon
) -> Self::DebugEpsilon

The epsilon used by a rel relative tolerance comparison, displayed when an assert fails.

Equivalent to self.debug_rmax_epsilon(self, other, max_diff), there is no need to reimplement this for your own types.

Loading content...