Trait float_eq::FloatEqUlpsTol

pub trait FloatEqUlpsTol {
    type UlpsTol: ?Sized;
}

Per-field tolerances for ULPs comparisons.

This trait establishes a one-to-one relation between an IEEE floating point type and a type whose fields are expected to be structurally identical but specified in ULPs. It is used by ULPs equality checks to specify per-field tolerances.

By convention, this type is named FooUlps for a given type Foo.

The UlpsTol type alias exists to simplify usage, for example UlpsTol<f32> is u32.

To implement this trait over a new type, see How to compare custom types.

Associated Types

type UlpsTol: ?Sized

A structurally identical type to Self, with fields recursively wrapped by UlpsTol.