Attribute Macro float_eq::derive_float_eq

#[derive_float_eq]

Helper for deriving the various float_eq traits.

By default, this will derive FloatEqUlpsTol, FloatEq, FloatEqDebugUlpsDiff and AssertFloatEq.

Attribute parameters are passed through to the #[float_eq(...)] attribute: ulps_tol is the name of the type generated for FloatEqUlpsTol and debug_ulps_diff the name of the type generated for FloatEqDebugUlpsDiff. Both are required.

If the optional all_tol parameter is provided, which names the type to be used by FloatEqAll::AllTol then FloatEqAll and AssertFloatEqAll are also derived.

See How to compare custom types for more information and example usage.