Trait float_eq::AssertFloatEqAll

pub trait AssertFloatEqAll<Rhs: ?Sized = Self>: FloatEqAll<Rhs> {
    type AllDebugEpsilon: Debug + FloatEqUlpsEpsilon;
    fn debug_abs_all_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::AllEpsilon
    ) -> Self::AllDebugEpsilon;

    fn debug_rmax_all_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::AllEpsilon
    ) -> Self::AllDebugEpsilon;

    fn debug_rmin_all_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::AllEpsilon
    ) -> Self::AllDebugEpsilon;

    fn debug_r1st_all_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::AllEpsilon
    ) -> Self::AllDebugEpsilon;

    fn debug_r2nd_all_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::AllEpsilon
    ) -> Self::AllDebugEpsilon;

    fn debug_ulps_all_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &UlpsEpsilon<Self::AllEpsilon>
    ) -> UlpsEpsilon<Self::AllDebugEpsilon>
    where
        UlpsEpsilon<Self::AllDebugEpsilon>: Sized;

    fn debug_rel_all_epsilon(
        &self, 
        other: &Rhs, 
        max_diff: &Self::AllEpsilon
    ) -> Self::AllDebugEpsilon { ... }
}

Debug context for when an assert using an all check fails.

This is used internally by the assert_float_eq! family of macros.

To implement this trait over a new type, see How to compare custom types.

Associated Types

type AllDebugEpsilon: Debug + FloatEqUlpsEpsilon

Displayed to the user when an assert fails, using fmt::Debug.

This should match the fields of the the most complex type in the comparison.

Loading content...
Required methods

fn debug_abs_all_epsilon(
&self,
other: &Rhs,
max_diff: &Self::AllEpsilon
) -> Self::AllDebugEpsilon

The epsilon used by an abs_all absolute tolerance comparison, displayed when an assert fails.

fn debug_rmax_all_epsilon(
&self,
other: &Rhs,
max_diff: &Self::AllEpsilon
) -> Self::AllDebugEpsilon

The epsilon used by an rmax_all relative tolerance comparison, displayed when an assert fails.

This should take into account the fact that the epsilon values are scaled based on the size of their inputs.

fn debug_rmin_all_epsilon(
&self,
other: &Rhs,
max_diff: &Self::AllEpsilon
) -> Self::AllDebugEpsilon

The epsilon used by an rmin_all relative tolerance comparison, displayed when an assert fails.

This should take into account the fact that the epsilon values are scaled based on the size of their inputs.

fn debug_r1st_all_epsilon(
&self,
other: &Rhs,
max_diff: &Self::AllEpsilon
) -> Self::AllDebugEpsilon

The epsilon used by an r1st_all relative tolerance comparison, displayed when an assert fails.

This should take into account the fact that the epsilon values are scaled based on the size of their inputs.

fn debug_r2nd_all_epsilon(
&self,
other: &Rhs,
max_diff: &Self::AllEpsilon
) -> Self::AllDebugEpsilon

The epsilon used by an r2nd_all relative tolerance comparison, displayed when an assert fails.

This should take into account the fact that the epsilon values are scaled based on the size of their inputs.

fn debug_ulps_all_epsilon(
&self,
other: &Rhs,
max_diff: &UlpsEpsilon<Self::AllEpsilon>
) -> UlpsEpsilon<Self::AllDebugEpsilon> where
UlpsEpsilon<Self::AllDebugEpsilon>: Sized,

The epsilon used by an ulps_all ULPs comparison, displayed when an assert fails.

Loading content...
Provided methods

fn debug_rel_all_epsilon(
&self,
other: &Rhs,
max_diff: &Self::AllEpsilon
) -> Self::AllDebugEpsilon

The epsilon used by a rel_all relative tolerance comparison, displayed when an assert fails.

Equivalent to self.debug_rmax_all_epsilon(self, other, max_diff), there is no need to reimplement this for your own types.

Loading content...