Trait float_eq::FloatEqDebugUlpsDiff

pub trait FloatEqDebugUlpsDiff {
    type DebugUlpsDiff;
}

Per-field results of ULPs based diff calculations.

This trait establishes a one-to-one relation between an IEEE floating point type and a type whose fields are expected to be structurally identical but specified as the result of calculating a diff in ULPs. It is used by testing and debugging tools to show the difference between two values on a per-field basis and is built for clarity, not runtime efficiency. The DebugUlpsDiff type alias exists to simplify usage, for example DebugUlpsDiff<f32> is Option<u32>. Usually, this type is named FooDebugUlpsDiff for a given type Foo.

To implement this trait over a new type, see How to compare custom types.

Associated Types

type DebugUlpsDiff

A structurally identical type to Self, with fields recursively wrapped by DebugUlpsDiff.

Loading content...